refactor(playground): share connect-channel side-effect routing
All five surfaces that connect an agent to a channel platform (chat
widget, publish dropdown, mobile menu, publish dialog, agent-channels
page) duplicated the same oauth/deep_link/immediate routing plus toast
error handling. Extract it into useConnectChannelAction(platformId,
{ onClose? }), which wraps useConnectChannel and exposes connect(agentId)
+ isConnecting. Net -58 lines.
