+++ packages/playground/src/domains/agent-builder/components/agent-builder-edit/publish-to-channel-button.tsx
@@ -1,4 +1,4 @@
-import { Button, DropdownMenu, StatusBadge, toast } from '@mastra/playground-ui';
+import { Button, DropdownMenu, StatusBadge } from '@mastra/playground-ui';
 import { ChevronDownIcon } from 'lucide-react';
 import { useState } from 'react';
 import { ChannelDialog } from './publish-channel-dialogs';
@@ -6,7 +6,7 @@ import { PlatformIcon } from '@/domains/agents/components/agent-channels/platfor
 import {
   useChannelInstallations,
   useChannelPlatforms,
-  useConnectChannel,
+  useConnectChannelAction,
 } from '@/domains/agents/hooks/use-channels';
 import type { ChannelInstallationInfo, ChannelPlatformInfo } from '@/domains/agents/hooks/use-channels';
 
@@ -70,7 +70,7 @@ interface PublishChannelMenuItemProps {
 function PublishChannelMenuItem({ platform, agentId, onSelect }: PublishChannelMenuItemProps) {
   const { data: installations = [] } = useChannelInstallations(platform.id, agentId);
   const installation = installations.find(i => i.status === 'active');
-  const { mutate: connect, isPending: isConnecting } = useConnectChannel(platform.id);
+  const { connect, isConnecting } = useConnectChannelAction(platform.id);
 
   // Slack-specific shortcut: when the platform is configured but not yet
   // connected, the dialog adds nothing — the user's intent ("connect Slack")
@@ -86,28 +86,7 @@ function PublishChannelMenuItem({ platform, agentId, onSelect }: PublishChannelM
       return;
     }
 
-    connect(
-      { agentId },
-      {
-        onSuccess: result => {
-          if (result.type === 'oauth') {
-            window.location.href = result.authorizationUrl;
-            return;
-          }
-          if (result.type === 'deep_link') {
-            const popup = window.open(result.url, '_blank', 'noopener,noreferrer');
-            if (!popup) {
-              toast.error('Popup blocked — please allow popups and try again');
-            }
-          }
-          // 'immediate' → installation list will be invalidated by the hook;
-          // no further UI action needed.
-        },
-        onError: (err: Error & { body?: { error?: string } }) => {
-          toast.error(err.body?.error || err.message || 'Failed to connect channel');
-        },
-      },
-    );
+    connect(agentId);
   };
 
   return (
